Decluttered server.js/  refactorings
