+++ config/passport.js
@@ -1,7 +1,6 @@
-const passport = require('passport')
-const jwt = require('jsonwebtoken')
-
-const { Strategy: JwtStrategy, ExtractJwt } = require('passport-jwt') 
+const passport = require('passport');
+const { Strategy: JwtStrategy, ExtractJwt } = require('passport-jwt')
+const User = require('../models/User')
 
 const passportOpts = {
     // Set Extraction method to pull it out from our header
@@ -14,6 +13,17 @@ passport.use(new JwtStrategy(
     passportOpts,
     (jwt_payload, done) => {
         console.log(jwt_payload);
+        User.findOne({ _id: jwt_payload._id }, (err, user) => {
+            if (err) {
+                return done(err, false)
+            }
+            console.log("user", user)
+            if (user) {
+                done(null, user)
+            } else {
+                done(null, false)
+            }
+        })
     }
 ))
